requirements.txt
requests
pandas
numpy
numba
tabulate
tqdm
ccxt
python-dotenv
websocket-client